까밋
까밋

+++ AllThatTip/src/main/java/com/kh/AllThatTrip/admin/model/vo/AdNotice.java
@@ -15,7 +15,7 @@ import lombok.ToString;
 @Getter
 public class AdNotice {
 
-	private int adNoticeNo;
+	private Long adNoticeNo;
 	private String importent;
 	private String adNoticeTitle;
 	private String adNoticeContent;
